requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
apscheduler==3.10.4
redis==5.0.8
httpx[test]==0.27.0
google-generativeai==0.8.6
anthropic>=0.50.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.24.0
respx==0.22.0
